Ban Evasion profiling
Keep track of who is ban evading, and what their alternate accounts/IP addresses are.

Provides routes for creating new profiles, linking accounts with profiles, linking ranges with profiles.
Provides routes for lookup by account/ip and allows for the removal of an account or IP from a ban evasion profile.

Moves everything into an /api/[category] directory.  A duplicate Comment route is left in place for now to provide backwards compatibility.
A seeder is added that will pre-populate the database with users already listed in the [mod ban evasion thread.](https://www.reddit.com/r/TagProModerators/comments/3zbhbg/ban_evasion_log/)

IP ranges are supported.  For the purposes of working with the TagPro Mod Tools, Only full octet ranges are used externally.
If the API is given "128.32.", it will store the range internally as 128.32.1.1 - 128.32.255.255.